backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.36
pandas>=2.2.0
numpy>=1.26.0
openai>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.6.0
httpx>=0.26.0
newsapi-python>=0.2.7
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
